Make binary-op dtype and typed-sink dtype known at plan time
Close the last two plan==exec divergences by giving each its single
authority, so the dtype Polars infers at planning time always equals the
dtype execution produces.

WS-6A — binary-op promotion (true division):
- Add BinaryOp::output_dtype(left, right) in view-buffer as the one authority
  for a binary op's result dtype, shared by planning and execution. Divide and
  Ratio now use true division: integer operands promote to float (f32, or f64
  when an operand is already f64). infer_dtype delegates to it, and execute()
  routes Divide/Ratio through the float path and guarantees the produced buffer
  carries exactly the authority dtype.
- Expose binary_output_dtype(op, left, right) over FFI; the lazy _binary_op
  planner now derives its dtype from BOTH operands via this hook instead of
  copying the left operand's, and propagates rank for downstream typed sinks.

WS-6B — typed sinks require a known element dtype:
- A list/array sink whose element dtype is "auto" (an image/blob source whose
  decoded dtype is only known at runtime) is rejected at plan time with a clear
  message asking for an explicit dtype, rather than silently guessing u8 and
  diverging from the realized dtype. List/array sources (dtype carried in the
  column schema) and binary/struct sinks are exempt. A defensive Rust guard
  refuses an unresolved "auto" reaching list/array schema resolution.

Hardening:
- The former xf*ail* divergences are now enforced tests; add a guard that fails
  (not skips) if a built plugin is missing an introspection hook, plus a
  two-input dtype-authority parity test over every binary op the API emits.
- Tests that previously relied on the u8 fallback now declare their dtype, and
  the divide/ratio reference tests assert true-division semantics.

Also normalizes pre-existing formatting flagged by current rustfmt.

+++ polars-cv/tests/test_lazy_composition_gaps.py
@@ -262,8 +262,9 @@ class TestBlendRatioExecution:
 
         result = df.select(out=expr1.ratio(expr2).sink("numpy"))
         arr = numpy_from_struct(result.row(0)[0])
+        # ratio is true division: u8 operands promote to f32.
         assert arr.shape == (30, 30, 3)
-        assert arr.dtype == np.uint8
+        assert arr.dtype == np.float32
 
 
 @plugin_required
